various changes

+++ include/task.inc.php
@@ -177,7 +177,8 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
         require_once XOOPS_ROOT_PATH . '/class/template.php';
         $xoopsTpl = new \XoopsTpl();
     }
-    // get template path
+
+    // get once template path
     $template_path = XNEWSLETTER_ROOT_PATH . '/language/' . $GLOBALS['xoopsConfig']['language'] . '/templates/';
     if (!is_dir($template_path)) {
         $template_path = XNEWSLETTER_ROOT_PATH . '/language/english/templates/';
@@ -376,22 +377,24 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
                 $letterTpl->assign('unsubscribe_link', XOOPS_URL . "/modules/xnewsletter/subscription.php?op=unsub&email={$recipient['address']}&actkey={$activationKey}");
                 $letterTpl->assign('unsubscribe_url', XOOPS_URL . "/modules/xnewsletter/subscription.php?op=unsub&email={$recipient['address']}&actkey={$activationKey}"); // new from v1.3
 
-                preg_match('/db:([0-9]*)/', $letterObj->getVar('letter_template'), $matches);
-                if (isset($matches[1])
-                    && ($templateObj = $helper->getHandler('Template')->get((int)$matches[1]))) {
-                    // get template from database
-                    $htmlBody = $letterTpl->fetchFromData($templateObj->getVar('template_content', 'n'));
-                } else {
-                    // get template from filesystem
-                    $template_path = XOOPS_ROOT_PATH . '/modules/xnewsletter/language/' . $GLOBALS['xoopsConfig']['language'] . '/templates/';
-                    if (!is_dir($template_path)) {
-                        $template_path = XOOPS_ROOT_PATH . '/modules/xnewsletter/language/english/templates/';
+                $templateObj = $helper->getHandler('Template')->get($letterObj->getVar('letter_templateid'));
+                if (is_object($templateObj)) {
+                    if ( (int)$templateObj->getVar('template_type') === _XNEWSLETTER_MAILINGLIST_TPL_CUSTOM_VAL) {
+                        // get template from database
+                        $htmlBody = $letterTpl->fetchFromData($templateObj->getVar('template_content', 'n'));
+                    } else {
+                        $template = $template_path . $templateObj->getVar('template_title') . '.tpl';
+                        $htmlBody = $letterTpl->fetch($template);
                     }
-                    $template = $template_path . $letterObj->getVar('letter_template') . '.tpl';
-                    $htmlBody = $letterTpl->fetch($template);
+                    try {
+                        $textBody = xnewsletter_html2text($htmlBody);
+                    }
+                    catch (Html2TextException $e) {
+                        $helper->addLog($e);
+                    }
+                } else {
+                    $htmlBody = _AM_XNEWSLETTER_TEMPLATE_ERR;
                 }
-                $textBody = xnewsletter_html2text($htmlBody); // new from v1.3
-                //$textBody = mb_convert_encoding($textBody, 'ISO-8859-1', _CHARSET); // "text/plain; charset=us-ascii" [http://www.w3.org/Protocols/rfc1341/7_1_Text.html]
 
                 $mail->addAddress($recipient['address'], $recipient['firstname'] . ' ' . $recipient['lastname']);
                 $mail->msgHTML($htmlBody); // $mail->Body = $htmlBody;
@@ -408,7 +411,7 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
                     if (0 == $subscr_id) {
                         $protocol_status        = _AM_XNEWSLETTER_SEND_SUCCESS_TEST . ' (' . $recipient['address'] . ')'; // old style
                         $protocol_status_str_id = _XNEWSLETTER_PROTOCOL_STATUS_OK_SEND_TEST; // new from v1.3
-                        $protocol_status_vars   = ['%recipient' => $recipient['address']]; // new from v1.3
+                        $protocol_status_vars   = ['recipient' => $recipient['address']]; // new from v1.3
                     } else {
                         $protocol_status        = _AM_XNEWSLETTER_SEND_SUCCESS; // old style
                         $protocol_status_str_id = _XNEWSLETTER_PROTOCOL_STATUS_OK_SEND; // new from v1.3
@@ -418,9 +421,9 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
                 } else {
                     $protocol_status        = _AM_XNEWSLETTER_FAILED . '-> ' . $mail->ErrorInfo; // old style
                     $protocol_status_str_id = _XNEWSLETTER_PROTOCOL_STATUS_ERROR_SEND; // new from v1.3
-                    $protocol_status_vars   = ['%error' => $mail->ErrorInfo]; // new from v1.3
+                    $protocol_status_vars   = ['error' => $mail->ErrorInfo]; // new from v1.3
 
-                    $protocol_success = false;
+                    $protocol_success = 0; //must be 0, because 'false' cause error when inserting protokol item
                     ++$count_err;
                 }
                 //create item in protocol for this email
@@ -429,10 +432,12 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
 
                 $mail->clearAddresses();
 
-                //delete item in table task
-                $sql_delete = "DELETE FROM {$xoopsDB->prefix('xnewsletter_task')}";
-                $sql_delete .= " WHERE `task_id`= {$recipient['task_id']}";
-                $result     = $xoopsDB->queryF($sql_delete);
+                //delete item in table task, if not from cron
+                if (0 < $uid) {
+                    $sql_delete = "DELETE FROM {$xoopsDB->prefix('xnewsletter_task')}";
+                    $sql_delete .= " WHERE `task_id`= {$recipient['task_id']}";
+                    $result = $xoopsDB->queryF($sql_delete);
+                }
 
                 $protocolObj = $helper->getHandler('Protocol')->create();
                 $protocolObj->setVar('protocol_letter_id', $letter_id);
@@ -446,7 +451,7 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
                 if ($helper->getHandler('Protocol')->insert($protocolObj)) {
                     // create protocol is ok
                 } else {
-                    echo $protocolObj->getHtmlErrors();
+                    echo $protocolObj->getHtmlErrors();die;
                 }
                 unset($protocolObj);
             }
@@ -471,7 +476,7 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
     if ($count_err > 0) {
         // IN PROGRESS
         $protocol_status  = xnewsletter_sprintf(_AM_XNEWSLETTER_SEND_ERROR_NUMBER, ['%e' => $count_err, '%t' => $count_total]);
-        $protocol_success = false;
+        $protocol_success = 0; //must be 0, because 'false' cause error when inserting protokol item
     } else {
         $protocol_success = true;
         if ($count_total > 0) {
@@ -486,7 +491,7 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
     $protocolObj->setVar('protocol_letter_id', $letter_id);
     $protocolObj->setVar('protocol_subscriber_id', 0);
     $protocolObj->setVar('protocol_status', $protocol_status);
-    $protocolObj->setVar('protocol_status_str_id', ''); // new from v1.3
+    $protocolObj->setVar('protocol_status_str_id', 0); // new from v1.3
     $protocolObj->setVar('protocol_status_vars', []); // new from v1.3
     $protocolObj->setVar('protocol_success', $protocol_success);
     $protocolObj->setVar('protocol_submitter', $uid);
@@ -494,7 +499,7 @@ function xnewsletter_executeTasks($xn_send_in_packages, $letter_id = 0)
     if ($helper->getHandler('Protocol')->insert($protocolObj)) {
         // create protocol is ok
     } else {
-        echo $protocolObj->getHtmlErrors();
+        echo $protocolObj->getHtmlErrors();die;
     }
     unset($protocolObj);
